Health declines when hungry or thirsty, added reactive combine extensions

+++ Assets/Statuses/HealthModel.cs
@@ -1,17 +1,18 @@
 ﻿using ReactiveCore;
+using ReactiveCore.Runtime;
 using Saving;
 
 namespace Statuses
 {
     public class HealthModel
     {
-        public ReactiveValue<int> Health { get; set; } = new(100);
+        public ReactiveValue<float> Health { get; set; } = new(100);
 
         public HealthModel()
         {
             if (ES3.KeyExists(SavegameConstants.HealthStatus))
             {
-                Health.Value = ES3.Load<int>(SavegameConstants.HealthStatus);
+                Health.Value = ES3.Load<float>(SavegameConstants.HealthStatus);
             }
             
             Health.Subscribe(value => ES3.Save(SavegameConstants.HealthStatus, value));
